chg: move endpoints into subfolder
- organise now since there'll probably be a few modules
